Modify CourseList class
Add check for update course if name is updated. This will cause a trouble because course_name is used as a key for the self.courses. If course_name is edited the prev course is deleted from the dictionary and a new key value pair with the new course name as a key and the updated course object as a value will be added to the dictionary.